refactored tests to remove duplication of test code
