+++ dbt-bigquery/tests/functional/functions/files.py
@@ -1,3 +1,7 @@
+from dbt.tests.adapter.functions.files import MASK_PII_JS  # noqa: F401
+from dbt.tests.adapter.functions.files import MY_JS_UDF  # noqa: F401
+from dbt.tests.adapter.functions.files import SUM_POSITIVE_JS  # noqa: F401
+
 MY_UDF_SQL = """
 price * 2
 """.strip()
@@ -63,11 +67,7 @@ functions:
 """.strip()
 
 
-# --- JavaScript UDF fixtures ---
-
-MY_JS_UDF = """
-return price * 2;
-""".strip()
+# --- JavaScript UDF YAML configs (BigQuery-specific types) ---
 
 MY_JS_UDF_YML = """
 functions:
@@ -92,17 +92,6 @@ functions:
       data_type: FLOAT64
 """
 
-MASK_PII_JS = """
-if (value === null || value === undefined) {
-    return null;
-}
-var masked = value.substring(0, 2);
-for (var i = 2; i < value.length; i++) {
-    masked += '*';
-}
-return masked;
-""".strip()
-
 MASK_PII_JS_YML = """
 functions:
   - name: mask_pii
@@ -115,21 +104,6 @@ functions:
 """
 
 # JS aggregate UDF — supported on BigQuery via CREATE AGGREGATE FUNCTION
-SUM_POSITIVE_JS = """
-export function initialState() {
-  return {sum: 0}
-}
-export function aggregate(state, x) {
-  if (x > 0) { state.sum += x; }
-}
-export function merge(state, partialState) {
-  state.sum += partialState.sum;
-}
-export function finalize(state) {
-  return state.sum;
-}
-""".strip()
-
 SUM_POSITIVE_JS_YML = """
 functions:
   - name: sum_positive
